+++ inst/tinytest/test-ab_property.R
@@ -24,6 +24,7 @@
 # ==================================================================== #
 
 expect_identical(ab_name("AMX", language = NULL), "Amoxicillin")
+expect_identical(ab_name("AMX", language = NULL, snake_case = TRUE), "amoxicillin")
 expect_identical(as.character(ab_atc("AMX")), "J01CA04")
 expect_identical(ab_cid("AMX"), as.integer(33613))